fix: move to @watchable/unpromise to dev deps
