+++ package.json
@@ -39,7 +39,6 @@
     "create:api": "node --experimental-strip-types scripts/create-api.ts"
   },
   "dependencies": {
-    "@watchable/unpromise": "^1.0.2",
     "ajv": "^8.17.1",
     "ajv-errors": "^3.0.0",
     "debug": "^4.4.0",
@@ -56,6 +55,7 @@
     "@types/debug": "^4.1.12",
     "@types/node": "^22.13.5",
     "@types/semver": "^7.7.0",
+    "@watchable/unpromise": "^1.0.2",
     "c8": "^10.1.3",
     "cleaner-spec-reporter": "^0.5.0",
     "cronometro": "^5.3.0",
